Add ExtraColorEditor, separate from edit page
