+++ src/components/layout/BaseColors/index.ts
@@ -1,6 +1,7 @@
 export * from './ColorListItem';
 export * from './DangerColorEditor';
 export * from './DebouncedColorPicker';
+export * from './ExtraColorEditor';
 export * from './NeutralColorEditor';
 export * from './PalettePreview';
 export * from './PrimaryColorEditor';
