+++ src/components/layout/OptionsDrawer/ExtraColorEditPage.tsx
@@ -1,25 +1,14 @@
 import {
   forwardRef,
-  useCallback,
   type ComponentPropsWithoutRef,
   type ElementRef,
 } from 'react';
-import { TrashIcon } from 'lucide-react';
 
 import { ColorEditPage } from './ColorEditPage';
-import { DebouncedColorPicker } from '@/components/layout/BaseColors';
-import { Input } from '@/components/ui/Input';
 import {
-  List,
-  ListItem,
-  ListItemIcon,
-  ListSubheader,
-} from '@/components/ui/List';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/Tabs';
-import {
-  ColorSuggestionButton,
-  ColorSuggestionsBox,
-} from '@/components/layout/ColorSuggestions';
+  ExtraColorEditor,
+  type ExtraColorEditorProps,
+} from '@/components/layout/BaseColors';
 import { useBaseColors } from '@/hooks/useBaseColors';
 import { useOptionsDrawer } from '@/store/useOptionsDrawer';
 import { cn } from '@/lib/utils';
@@ -29,44 +18,24 @@ import {
 } from '@/constants';
 import type { GeneralColorSuggestion } from '@/types/defaultSuggestions';
 
-type ExtraColorEditPageProps = ComponentPropsWithoutRef<'div'> & {
-  index: number;
-};
+type ExtraColorEditPageProps = ComponentPropsWithoutRef<'div'> &
+  ExtraColorEditorProps;
 
 export const ExtraColorEditPage = forwardRef<
   ElementRef<typeof ColorEditPage>,
   ExtraColorEditPageProps
 >(({ index, className, ...props }, ref) => {
   const { primary, extras: baseExtras } = useBaseColors(),
-    {
-      extras: editedExtras,
-      removeExtraColor,
-      renameExtraColor,
-      setExtraColor,
-    } = useOptionsDrawer();
+    { extras: drawerExtras } = useOptionsDrawer();
 
-  const { name, value } = editedExtras?.[index] || baseExtras[index];
+  const { name, value } = drawerExtras?.[index] || baseExtras[index];
   const title = name || `Extra ${index + 1}`;
 
-  const colorIsSuggestion = generalColorSuggestionNames.includes(value),
-    colorIsCustom = !generalColorSuggestionNames.includes(value);
+  const colorIsSuggestion = generalColorSuggestionNames.includes(value);
 
   const themeValue = colorIsSuggestion
     ? generalColorSuggestions[value as GeneralColorSuggestion]?.(primary)
     : value;
-
-  const handleValueChange = useCallback(
-    (newValue: string) => {
-      if (value === newValue) return;
-      setExtraColor(index, newValue);
-    },
-    [index, setExtraColor, value]
-  );
-
-  const handleRemove = useCallback(() => {
-    setTimeout(() => removeExtraColor(index), 90);
-    window.history.back();
-  }, [index, removeExtraColor]);
 
   return (
     <ColorEditPage
@@ -80,55 +49,7 @@ export const ExtraColorEditPage = forwardRef<
         className
       )}
     >
-      <Input
-        label='Name'
-        value={name || ''}
-        onChange={e => renameExtraColor(index, e.target.value)}
-      />
-      <Tabs
-        defaultValue={colorIsCustom ? 'custom' : 'suggestions'}
-        className='mt-2'
-      >
-        <TabsList className='grid w-full grid-cols-2'>
-          <TabsTrigger value='suggestions'>Suggestions</TabsTrigger>
-          <TabsTrigger value='custom'>Custom</TabsTrigger>
-        </TabsList>
-        <TabsContent value='suggestions'>
-          <ColorSuggestionsBox
-            value={value || 'auto'}
-            onValueChange={handleValueChange}
-            className='p-0 [&>span]:col-span-full [&>span]:px-1'
-          >
-            <ListSubheader>Extra color suggestions</ListSubheader>
-            {Object.entries(generalColorSuggestions).map(
-              ([value, variantFn]) => (
-                <ColorSuggestionButton
-                  key={value}
-                  value={value}
-                  color={variantFn(primary)}
-                />
-              )
-            )}
-          </ColorSuggestionsBox>
-        </TabsContent>
-        <TabsContent value='custom'>
-          <DebouncedColorPicker
-            initialValue={themeValue}
-            onChange={handleValueChange}
-          />
-        </TabsContent>
-      </Tabs>
-      <List
-        className='absolute inset-x-0 bottom-0 z-10 justify-center
-bg-background p-2 dark:bg-card'
-      >
-        <ListItem onClick={handleRemove}>
-          <ListItemIcon>
-            <TrashIcon />
-          </ListItemIcon>
-          Remove
-        </ListItem>
-      </List>
+      <ExtraColorEditor index={index} />
     </ColorEditPage>
   );
 });
